backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0
psycopg2-binary==2.9.7
redis==5.0.1
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
cryptography==41.0.7
celery==5.3.4
gunicorn==21.2.0
whitenoise==6.6.0